Unify WasmVersion into concordium_contracts_common
